Add files via upload
This repository contains my submission for the internship program application at Maybank. The assignment involved implementing a batch job to consume a text file and developing a RESTful API for retrieving and updating records. The repository includes documentation, diagrams, and descriptive explanations related to the assignment.

Feel free to explore the repository to review my solution to the assignment.